Rename our core module to `_runtime`
