+++ tractor/_spawn.py
@@ -42,7 +42,7 @@ from ._state import (
 
 from .log import get_logger
 from ._portal import Portal
-from ._actor import Actor
+from ._runtime import Actor
 from ._entry import _mp_main
 from ._exceptions import ActorFailure
